gbp pq import/export and capture date
Minimize the diffs for future patch modifications by standardizing the
format of the patch queue on git-buildpackage.  This operation should
not change the content of the patch in any way.

Furthermore, I have manually adjusted the Date: header to be precisely
the date of the last commit modifying the patch file.

Signed-off-by: Antonio Enrico Russo <aerusso@aerusso.net>